refactor: Header Dropdown Menu 컴포넌트 style 코드 분리 #7
